fix: useModalForm & useStepsForm return type
