+++ .changeset/tall-garlics-fly.md
@@ -1,0 +1,5 @@
+---
+"@pankod/refine-antd": patch
+---
+
+Fixed `useModalForm` & `useStepsForm` return type
