+++ packages/antd/src/types/sunflower.d.ts
@@ -1,46 +1,11 @@
 import { FormInstance, FormProps, ModalProps } from "../components/antd";
 import { UseFormConfig, UseModalFormConfig } from "sunflower-antd";
-import { BaseKey } from "@pankod/refine-core";
-
 export interface UseStepsFormConfig
     extends Omit<UseFormConfig, "defaultFormValues"> {
     defaultCurrent?: number;
     total?: number;
     isBackValidate?: boolean;
 }
-
-export type useModalFormFromSFReturnType<TData, TVariables> = {
-    form: FormInstance<TVariables>;
-    visible: boolean;
-    show: (id?: BaseKey) => void;
-    close: () => void;
-    modalProps: ModalProps;
-    formProps: FormProps<TVariables>;
-    formLoading: boolean;
-    defaultFormValuesLoading: boolean;
-    formValues: {};
-    initialValues: {};
-    formResult: undefined;
-    submit: (values?: TVariables) => Promise<TData>;
-};
-
-export type useStepsFormFromSFReturnType<TData, TVariables> = {
-    current: number;
-    gotoStep: (step: number) => void;
-    stepsProps: {
-        current: number;
-        onChange: (currentStep: number) => void;
-    };
-    formProps: FormProps<TVariables>;
-    formLoading: boolean;
-    defaultFormValuesLoading: boolean;
-    formValues: {};
-    initialValues: {};
-    formResult: undefined;
-    form: FormInstance<TVariables>;
-    submit: (values?: TVariables) => Promise<TData>;
-};
-
 declare module "sunflower-antd" {
     export const useStepsForm: <TData, TVariables>(
         config: UseStepsFormConfig,
